[UIK-3330][docs] Added role img to Flags in input phone (#2037)
<!--- Provide a general summary of your changes in the Title above -->

## Motivation and Context

Added `role=img` to InputPhone examples (2nd and 3d), so that flag in
input/select trigger is announced as full country name.

Also edited aria-labels in the 2nd example, so that SR reads when
focusing the input:
"Phone number, United States, ..."

(Flags in the Select Menu aren't announced, this is by design.)

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue).
- [ ] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected).
- [x] Nice improve.

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] I have updated the documentation accordingly or it's not required.
- [ ] Unit tests are not broken.
- [ ] I have added changelog note to corresponding `CHANGELOG.md` file
with planned publish date.
- [ ] I have added new unit tests on added of fixed functionality.