[slider] fix issue with value > max not clamping or decreasing on key (#2232)
<!--- Provide a general summary of your changes in the Title above -->

## Motivation and Context

When a user entered a value greater than the slider's maximum via the
input field, the slider knob did not clamp to the maximum value.
Additionally, when the knob was focused, pressing the Left Arrow key did
not decrease the value as expected.

This fix ensures the value is clamped to the maximum if the input
exceeds it, and that the knob responds correctly to arrow key
navigation.

## How has this been tested?

It's been tested manually.

## Screenshots (if appropriate):

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Bug fix (non-breaking change which fixes an issue).
- [ ] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected).
- [ ] Nice improve.

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] I have updated the documentation accordingly or it's not required.
- [x] Unit tests are not broken.
- [x] I have added changelog note to corresponding `CHANGELOG.md` file
with planned publish date.
- [ ] I have added new unit tests on added of fixed functionality.

---------

Co-authored-by: Valeryia Zimnitskaya <valeryia.zimnitskaya@semrush.com>