[UIK-3699][d3-chart] show correct data in tooltip on hover for CompactHorizontalBar (#2316)
<!--- Provide a general summary of your changes in the Title above -->

## Motivation and Context

In the Compact Horizontal Bar chart, the tooltip always displayed data
from the first bar, regardless of which bar was hovered.
This fix ensures that the tooltip correctly reflects the data of the
hovered bar.

## How has this been tested?

It's been tested manually (visually)

## Screenshots (if appropriate):
<img width="595" alt="Screenshot 2025-07-07 at 10 32 35"
src="https://github.com/user-attachments/assets/7850b473-bec7-4caf-9f35-1380d743d5ff"
/>

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Bug fix (non-breaking change which fixes an issue).
- [ ] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected).
- [ ] Nice improve.

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] I have updated the documentation accordingly.
- [x] I have added changelog note to corresponding `CHANGELOG.md` file
with planned publish date.
- [ ] I have added new tests on added of fixed functionality.