[UIK-4032] added canva icon (#2384)
<!--- Provide a general summary of your changes in the Title above -->

## Motivation and Context

Added a new Canva icon.

One caveat: I left the white fill for the letter "C" because their
guidelines say that only white or their gradient are allowed as fills
for the letter.

And their only example of an icon with just the letter "C" is with a
circle on the background, in other words it seems they don't allow just
letter "C" without the circle.

So I decided to add it as a monochromatic icon, so that the circle can
be colored, but the letter C is always white.

<img width="933" height="423" alt="imagen"
src="https://github.com/user-attachments/assets/158aec8c-3877-4615-b8b4-cabdc37530d7"
/>

<img width="826" height="647" alt="imagen"
src="https://github.com/user-attachments/assets/7e943fc6-f90e-4bcd-8929-9ec8f319a8f9"
/>

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue).
- [x] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected).
- [ ] Nice improve.

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] I have updated the documentation accordingly.
- [x] I have added changelog note to corresponding `CHANGELOG.md` file
with planned publish date.
- [ ] I have added new tests on added of fixed functionality.